get vfprintf compiled

+++ linux/src/drivers/block/ide-cd.h
@@ -528,7 +528,7 @@ struct atapi_mechstat_header {
 
 	byte     curlba[3];
 	byte     nslots;
-	__u8 short slot_tablelen;
+	short slot_tablelen;
 };
